+++ babel.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   presets: ['codecademy', '@babel/preset-typescript'],
   plugins: [
+    'macros',
     'react-anonymous-display-name',
     [
       '@emotion',
@@ -8,14 +9,8 @@ module.exports = {
         sourceMap: true,
         autoLabel: 'always',
         labelFormat: '[local]',
-        cssPropOptimization: true,
       },
     ],
   ],
   ignore: [],
-  env: {
-    test: {
-      plugins: ['require-context-hook'],
-    },
-  },
 };