+++ packages/gamut-labs/src/brand/Header/HeaderContainer/index.tsx
@@ -6,6 +6,9 @@ import styles from './styles.module.scss';
 
 export type HeaderContainerProps = {
   className?: string;
+  /**
+   * Pass ReactNode into any or each desired positions
+   */
   sections?: {
     after?: React.ReactNode;
     left?: React.ReactNode;