+++ packages/gamut-labs/src/landingPage/PagePrefooter.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import React from 'react';
 
 import { CTA, Description, Title } from './';
-import { BaseProps } from './types';
+import { BaseProps, DarkModeProps } from './types';
 
 const FlexContainer = styled(Container)`
   ${mediaQueries.sm} {
@@ -23,27 +23,31 @@ const StyledCTA = styled(CTA)`
   }
 `;
 
-export const PagePrefooter: React.FC<BaseProps> = ({
+export const PagePrefooter: React.FC<BaseProps & DarkModeProps> = ({
   title,
   desc,
   cta,
+  onAnchorClick,
+  mode,
   testId,
 }) => {
-  const SectionTitle = title && <Title>{title}</Title>;
-  const Desc = desc && <Description text={desc} />;
+  const SectionTitle = title && <Title mode={mode}>{title}</Title>;
+  const Desc = desc && (
+    <Description text={desc} onAnchorClick={onAnchorClick} mode={mode} />
+  );
 
   return cta ? (
-    <FlexContainer nowrap column justify="spaceBetween">
+    <FlexContainer data-testid={testId} nowrap column justify="spaceBetween">
       <FlexContent>
         {SectionTitle}
         {Desc}
       </FlexContent>
-      <StyledCTA href={cta.href} onCtaButtonClick={cta.onClick}>
+      <StyledCTA href={cta.href} onCtaButtonClick={cta.onClick} mode={mode}>
         {cta.text}
       </StyledCTA>
     </FlexContainer>
   ) : (
-    <div>
+    <div data-testid={testId}>
       {SectionTitle}
       {Desc}
     </div>