+++ packages/gamut-labs/src/landingPage/types.tsx
@@ -1,9 +1,17 @@
+import { HandlerProps } from '@codecademy/gamut-system';
+
+import { darkModeVariants } from './variants';
+
 export type BaseProps = {
   title?: string;
   /**
-   * Main body text (can include markdown)
+   * Body text as markdown
    */
   desc?: string;
+  /**
+   * Callback when a markdown anchor tag is clicked
+   */
+  onAnchorClick?: (event: React.MouseEvent<HTMLAnchorElement>) => void;
   /**
    * text: button text
    *
@@ -18,3 +26,5 @@ export type BaseProps = {
   };
   testId?: string;
 };
+
+export type DarkModeProps = HandlerProps<typeof darkModeVariants>;