+++ packages/styleguide/.storybook/components/PropsTable/Elements.tsx
@@ -1,11 +1,6 @@
-import styled from '@emotion/styled';
-import {
-  boxShadow,
-  colors,
-  fontAccent,
-  fontMonospace,
-  spacing,
-} from '@codecademy/gamut-styles';
+import { styled } from '@storybook/theming';
+import { boxShadow, spacing } from '@codecademy/gamut-styles';
+import { css } from '@emotion/react';
 
 export const PropGroupTooltip = styled.ul`
   display: none;
@@ -14,8 +9,8 @@ export const PropGroupTooltip = styled.ul`
   left: 50%;
   transform: translate(-50%, 0);
   padding: ${spacing[8]} ${spacing[12]};
-  background-color: ${colors['gray-100']};
-  border: 1px solid ${colors['gray-200']};
+  background-color: ${({ theme }) => theme.color.light};
+  border: 1px solid ${({ theme }) => theme.color.medium};
   list-style: none;
   margin: 0;
   border-radius: 6px;
@@ -25,24 +20,24 @@ export const PropGroupTooltip = styled.ul`
 export const PropItem = styled.li`
   margin: ${spacing[8]} 0;
   font-size: 11px;
-  font-family: ${fontMonospace};
+  font-family: ${({ theme }) => theme.typography.fonts.mono};
   text-align: center;
   text-transform: none;
-  color: ${colors['gray-700']};
+  color: ${({ theme }) => theme.color.dark};
 `;
 
 export const HeaderColumn = styled.div``;
 
-export const PropGroupTag = styled.span`
+export const PropGroupTag = styled.span<{ active?: boolean }>`
+  user-select: none;
   position: relative;
   display: inline-block;
   padding: ${spacing[4]} ${spacing[8]};
   font-size: 11px;
-  font-family: ${fontAccent};
+  font-family: ${({ theme }) => theme.typography.fonts.mono};
+  font-weight: 700;
   margin: ${spacing[4]};
   margin-top: 0;
-  color: ${colors['gray-600']};
-  border: 1px solid ${colors['gray-400']};
   border-radius: 4px;
   text-transform: uppercase;
   cursor: help;
@@ -52,29 +47,35 @@ export const PropGroupTag = styled.span`
   }
 
   &:hover {
-    ${PropGroupTooltip} {
+    > * {
       display: block;
     }
   }
-`;
 
-export const Header = styled.div`
-  display: grid;
-  grid-template-columns: max-content 1fr max-content;
+  ${({ active, theme }) =>
+    active
+      ? css`
+          color: ${theme.color.secondary};
+          border: 1px solid ${theme.color.secondary};
+        `
+      : css`
+          color: ${theme.color.dark};
+          border: 1px solid ${theme.color.mediumdark};
+        `}
 `;
 
 export const Title = styled.p`
   display: inline-block;
+  font-weight: 700;
   margin: 0;
-  margin-right: ${spacing[8]};
 `;
 
 export const ToggleContainer = styled.div`
   display: flex;
-  align-items: center;
+  align-items: flex-start;
 `;
 
-export const ToggleLabel = styled.span`
+export const ToggleLabel = styled.label`
   margin-right: 4px;
   font-size: 14px;
   line-height: 2;