+++ packages/styleguide/.storybook/manager.ts
@@ -1,3 +1,6 @@
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-a11y/register';
+import { addons } from '@storybook/addons';
+import { theme } from './theme';
+
+addons.setConfig({
+  theme,
+});